Rename old Uploadex module to Context and UploadexUse to Uploadex
